refactor: 日付フォーマット関数の重複を解消
6つのファイルに重複していた日付フォーマット関数を統一。
`src/utils/date.ts`に以下の関数を集約：
- toDateInputValue: input[type="date"]用のYYYY-MM-DD形式
- formatDeadlineForDisplay: 表示用のM/DD(曜日)形式
- toYmd: 既存のnull対応版（保持）

変更ファイル：
- frontend/src/utils/date.ts: 統一関数を追加
- frontend/src/features/tasks/inline/InlineTaskRow.tsx: 重複削除
- frontend/src/features/tasks/inline/TaskRowDisplay.tsx: 重複削除
- frontend/src/features/tasks/inline/useTaskEditing.ts: 重複削除
- frontend/src/features/tasks/nest.ts: toDateInput→toDateInputValueに統一
- frontend/src/features/priority/PriorityTasksPanel.tsx: fmtDeadline→formatDeadlineForDisplayに統一

効果：
- コード重複を約50行削除
- 単一の責任（Single Source of Truth）を実現
- 保守性とテスト容易性が向上
